Failed attempt at making backup handlers run first
The logic here just doesn't work. You need to know that there's not an
acceptable symmetric solution before you pass stuff off to the backup
handler otherwise the handler will come up with the wrong resolution.

For example, with circular movement, one of the units in the movement
could bounce which would produce a symmetric solution (same outcome for
both guesses) but doing things this way would result in a successful
resolution anyway.