Updated constructor signature on Engine - now depends directly on data
- The Engine is now serving as a facade into all of this Gavel stuff.
Doing things this way makes initialization easier for the client and it
also makes it easier to pass around the needed dependencies (you can
pass around the engine instead of having to pass around the engine, the
board, and the pathfinder). The theoretical downside to this would be
that the engine is overly coupled with the concrete implementations of
Board, etc. This should be fairly easy to work around though (allow it
to be overridden) so that we can still allow custom implementations.